+++ src/features/ContentOverviewHero/ui/ContentOverviewActionButton/ContentOverviewActionButtons.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { fn } from 'storybook/test';
 
-import { ContentOverviewActionButtons } from '@/features/ContentOverviewHero/ui/ContentOverviewActionButtons';
+import { ContentOverviewActionButtons } from './ContentOverviewActionButtons';
 
 const meta = {
   title: 'Features/ContentOverviewHero/ContentOverviewActionButtons',
