+++ src/features/ContentOverviewHero/ui/ContentOverviewActionButton/ContentOverviewActionButtons.tsx
@@ -1,7 +1,7 @@
 import { MapPin, Map } from 'lucide-react';
 import { IconButton } from '@/shared/ui';
-import type { ContentOverviewActionButtonsProps } from '../model/types';
-import { messages } from '../model/messages';
+import type { ContentOverviewActionButtonsProps } from '../../model/types';
+import { messages } from '../../model/messages';
 
 export function ContentOverviewActionButtons({
   onLocationViewClick,
